sigi_backend/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
mysqlclient==2.2.4
Pillow==10.2.0
openpyxl==3.1.2
reportlab==4.1.0
